Refactor adjusting functions to a new module and update references across the codebase
