+++ libs/functions/__init__.py
@@ -5,5 +5,6 @@
 - :doc:`libs.functions.search`: DB検索
 - :doc:`libs.functions.message`: メッセージ
 - :doc:`libs.functions.compose`: メッセージ生成
+- :doc:`libs.functions.adjusting`: 表・桁数・調整
 - :doc:`libs.functions.tools`: 外部ツール用モジュール
 """
